separate build and encode operations for curam atp and mec check
